Stop ignoring generated files
While it is possible that people have inkscape and are able to generate
the png and pdf files themselves, having one correctly generated is
useful, if we want to quickly download it, or to compare a "validated"
(commited) result with one self-generated.

As it happens, installing inkscape and using the Makefile doesn't
generate the png file correctly, in one test machine.